sub-agents: trust the model, drop the iteration cap (#1127)
Per DESIGN.md P3 ("Build for the world six months from now"), removes the
hardcoded 20-iteration ceiling on sub-agent loops. Termination is driven
by the model (clean stop, no tool calls), `LoopDetector` (consecutive
identical calls -> feedback -> hard stop), parent cancellation, or
context exhaustion -- the same set of mechanisms Codex and Zed rely on.

Audit of reference codebases:
- Codex: no iteration cap, no heartbeat, `Errored(String)` on failure
- Zed: no iteration cap, no loop detection, `enum CompletionError {
  MaxTokens, Refusal, Other }` only

We have *more* safety than either (LoopDetector + sandbox + worktree
budget + structural-failure markers); the cap was redundant scaffolding
for a model weakness P3 explicitly tells us not to compensate for.

Changes (#1110):
- delete `MAX_SUB_AGENT_ITERATIONS = 20` and the bounded loop
- delete `iteration_cap_marker` + its `b18_tests` mod
- widen `AgentStatus::Running.iter` u8 -> u32 (a meandering weak model
  could plausibly exceed `u8::MAX` before context exhaustion; silent
  wraparound would mislead the TUI)
- drop the "/20" denominator from TUI iter rendering
- add "consider switching to a stronger model" hint to the LoopDetector
  hard-stop message (the one place we have a smoking-gun signal)
- `InvokeAgent` description gets one line steering read-only work to
  `explore` (faster, cheaper, no isolated workspace) and writes to `task`

Changes (#1106):
- bump `WAIT_TASK_DEFAULT_TIMEOUT_SECS` 30 -> 60 (sub-agent inference
  rounds take real time; default should mean "I'm willing to wait this
  long", not "check back quickly")
- description nudge: prefer 120-300 s for sub-agent waits, call sparingly

Net diff: +72 / -203 lines.

Tests: full workspace `cargo test --lib --tests` passes (1934 lib +
integration tests, 0 failures); `cargo clippy --all-targets -D warnings`
clean; `cargo fmt` clean.

Closes #1110
Closes #1106

+++ koda-core/src/bg_agent.rs
@@ -70,9 +70,11 @@ use crate::engine::EngineEvent;
 /// when execution actually starts and to one of the terminal variants
 /// (`Completed`, `Errored`, `Cancelled`) when it finishes.
 ///
-/// `Running.iter` reflects the current inference iteration (1..=20).
-/// Background agents emit live updates via Layer 4 (#1058); `0` is
-/// the entry-point placeholder before the first iteration fires.
+/// `Running.iter` reflects the current inference iteration. Background
+/// agents emit live updates via Layer 4 (#1058); `0` is the entry-point
+/// placeholder before the first iteration fires. There is no upper bound
+/// (#1110 removed the sub-agent cap) — the counter widens to `u32` so
+/// long-running agents can't silently wrap.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(tag = "kind", rename_all = "snake_case")]
 pub enum AgentStatus {
@@ -81,11 +83,13 @@ pub enum AgentStatus {
     /// Actively executing. `iter` is the current inference iteration
     /// (1..=20); `0` means "started, no iter info yet" (Layer 0 default).
     Running {
-        /// Current inference iteration (1..=20). `0` is the
-        /// entry-point placeholder emitted before the first iteration
-        /// in `run_bg_agent`; background agents update this live
-        /// (Layer 4, #1058).
-        iter: u8,
+        /// Current inference iteration (1..). `0` is the entry-point
+        /// placeholder emitted before the first iteration in
+        /// `run_bg_agent`; background agents update this live
+        /// (Layer 4, #1058). `u32` because #1110 removed the sub-agent
+        /// iteration cap and a meandering weak model could plausibly
+        /// exceed `u8::MAX` before context exhaustion.
+        iter: u32,
     },
     /// User or parent fired the cancel token. Terminal.
     Cancelled,
